clean: Fixing some new warnings produced by GCC9
Fixing some warnings that are showing up with newer GCC 9.2 and were not previously seen.
The warnings fixed in here are based on strncpy erroring based off of using a size that is not necessarily the destination size and/or is based off the source size which produces this warning. This likely would not have been an issue though since the M_Min macro was being used, but the warning is fixed anyways.