feat: Changing permissions error to warning
Changing from an error to a warning when not run with the correct permissions.
This was done at request to minimize permission requirements in Linux.
In Linux, a user can be added to the "disk" group which allows sending commands without needing sudo or root which is why this was requested.