feat: Fixing many MinGW warnings and changing how -F is parsed
Fixing numerous MinGW warnings found with -Wextra in the utilities.

Also changing how the scan flags (-F) is parsed to using a structure and a function. This makes it easier to maintain than the ugly macro that was in use previously.
Due to this change, all utilities have received a major version update as this will not be backwards compatible.
Basically, if you want multiple -F arguments, instead of doing "-F ata usbInterface", you would now do "-F ata -F usbInterface"