feat: Adding depopulate/repopulate commands to openSeaChest_Format
Adding the storage element depopulation and new repopulation commands to SeaChest_Format. These make more sense here than in the Erase tool as these are reconfiguration commands that internally reformat the drive for these changes and are not necessarily just to erase data. They will erase data, but that is not really their purpose.