doc: Removing use of --single-branch from MSYS2 instructions
removing example use of --single-branch from this part of the readme in relation to issue #44 discussion.

Instead of it being part of the instructions or walk though, it has been moved to a note that if it was used, how to undo it and update the submodules that were pulled with this. This should reduce confusion from these instructions.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>