bug: Improving IDD progress check and other language
Adding some language to help inform users about trouble pinging the drive for IDD progress or abort an IDD in the first 2 minutes. This is not possible due to how the drive performs the IDD test, so additional messaging has been added to help indicate this.
If these are attempted at this time, it may cause an OS to issue a reset and stop the IDD whether wanted or not. It may also cause the software to appear to have hung until the drive is finally ready again.

More improvements may be possible in the future.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>