bug: Changing fast format to stay busy until complete
Changing fast format options to hold the bus busy until the format is complete in order to make sure the drive completes the format as expected without interruptions from the OS or the user.
This was added after observing that drives doing fast format report 100% right away, however this is confusing as the sense data still shows that the drive is busy with a format. Setting it to hold busy makes it more clear when the drive is actually done and ready.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>