doc: Updating and reorganizing markdown documentation files
Updating the order of the information in the main readme.md document to help it flow better.
Split build instructions out to building.md and consolidated msys2 instructions into this file as well.
Adding links to security.md in contributing and readme.md files.
Adding commit template that was not previously checked in.

Lots of other little updates based on a markdown linter installed in VSCode to make them free of errors as best as possible.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>