feat: Automatically running Seagate quick format if setSectorSize fails
Automatically calling the Seagate quick format in case setSectorSize fails to complete properly. This is being done to reduce the chance of the drive being in a bad state when this happens.

It seems unlikely that this will happen as other changes to lock the device (switch to blocking io) will also prevent this in most cases, but the drive should not be in a bad state after this command, whether it passed or failed.
Once quick format has completed, it should be in a better state to retry the sector size change if it is not correct.

[Seagate/ToolBin#16]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>