feat: Adding new option for ignoring the final download segment
Adding a new command line option for ignoring the final status of a firmware update. This is a workaround for problems when updating older drives. Since these haven't received updates in years and the "automatic" approach that was in the code before kept causing problems, this has been moved to a separate option.
Additionally, the other FWDL structure updates are pulled in to help manage versions and sizes of structures to try and improve compatibility.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>