bug: Fixing a bug in reading a file for changing SCSI mode pages
Fixing a bug in how successful read of a file was judged when trying to change SCSI mode pages. The issue was attempting to read and compare to a size greater than the file's actual length. This was due to adding an extra space for a NULL in the memory, but not taking that into account during the file read.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>