feat: Better handling of Linux Libc license to handle musl vs glibc
Moved some code around and changed how the libc license is detected since some distributions may use musl instead of glibc.
If glibc is detected, it is used. Otherwise a preprocessor definition of USING_MUSL_LIBC can be used to show the musl license. If neither is detected, a warning is emitted.

[#57]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>