feat: Minor testing update and added detection of TPID hack
Minor changes to the tests and outputs.
Most notable change is the addition of TPID hack which says that TPSIU can be used for transfer length with identify device commands. This may need further expansion to search for other commands it works on as the official hack in opensea-transport is listed as "limitedTPSIU" in case it needs to be flexible in the future...probably won't but it won't hurt.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>