bug: Fixing a bug in the --setSCSIMP option that caused a crash
Fixing a bug that would crash openSeaChest_Configure inside the --setSCSIMP option.
The issue was that the function scsi_Set_Mode_Page takes a buffer that should contain ONLY the mode page data, no headers or data block descriptors, but in the case of providing the byte offset, bit, value to make a change, it was providing the header which caused a mismatch in what data was supposed to be sent to the drive and caused a heap corruption.

The solution was to take into account the differences in how this data was read from the function that reads the mode page. This meant adjust the pointer to the correct offset so that the beginning of the mode page is passed as expected.
This should work on current and old drives as it checks whether mode sense 10 or mode sense 6 was used and it checks for original SCSI for a data block descriptor in the output (although anything that old probably won't be used anymore).

[Fixes #59]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>