make: initial support for meson build system
The meson build system is a build system with many advantages over others, and one I consider to be the best. It is one of the fastest build systems, has an easy to understand, and extensible syntax (for example a target can be defined to automatically build docs). It supports all platforms the other build systems do (Linux, Windows, FreeBSD) and is easy to use when cross compiling via a cross file (such as with MingW). It can also programmatically generate the version, instead of needing to edit a header file every time. On my Linux system, meson (configure) took 0.65s and ninja (build) took 16.62s, while the make-based build system took 30.22s.
See https://mesonbuild.com/Overview.html and https://mesonbuild.com/Comparisons.html for me info on the meson build system.
Meson requires all subprojects to be in the subprojects directory. It unfortunately does not support symlinks in the subprojects directory, so I updated existing submodules to be symlinks to submodules in the subprojects directory. See https://github.com/mesonbuild/meson/issues/8827 for more info.
Meson also natively supports CMake subprojects (such as wingetopt) and supports configurable options, such as whether TCG should be used or if the system's libc is musl.
Note: existing build systems can still be used.

[#2]

Signed-off-by: Reagan Bohan <xbjfk.github@gmail.com>