lib: Pulling in updates for better SSD detection on new USB products
Pulling in library changes to fix detection of SSDs for some new USB SSDs since they report differently than we've seen in the past.
Most of the changes are generic enough to work with anything similarly attached, but there are some specific changes based on what we've been able to test.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>