lib: Pulling in updates for analyzed smart attribute output
Pulling in an update related to how some smart attributes were being output. The function was reporting a warning on the GCC compiler in Illumos for potentially accessing a negative offset, which would be bad. Swapping to M_BytesToXByte macros made more sense here and made the code match other parts of the same function.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>