make: auto gnu99 mode for older gcc
Automatically detecting GCC 5.5.x and older to set gnu99 mode. Older GCC defaulted to gnu89 which does not compile as this project needs some of C99 features to compile.
This may need fine tuning in the future as I can only find the 5.5.0 docs online and not sure the exact version that the default C mode changed in.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>