feat: Switching to using wingetopt for getopt_long implementation
Switching to using wingetopt for a common getopt_long implementation that works on various platforms.
Wingetopt is now poorly named, but that is ok for now. :)
AIX needed a getopt_long implementation to use and this made the most sense to get working in that environment. I have tested that it runs in FreeBSD, Solaris, & MinGW builds with Meson and have not seen any issues.

It would be possible to make a flag for this, but for now this seemed easy enough and works where it needs to.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>