doc: Adding examples for using most, if not all options, in the tools
Adding examples for how to use the various options and combinations of options to each tool. There is room to add more, but this should cover the most common questions related to "how do I do xyz".
This follows on to an earlier commit that added more examples to openSeaChest_Erase. All of the different tools need these kinds of examples.

[Seagate/ToolBin#27]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>