feat: Adding new option for nvme Log Length to assist pulling logs generically
Adding a new option to assist pulling NVMe logs with the generic log pull. This is especially helpful for vendor unique logs that need to be pulled from the device.
There is not a standardized way to figure out the length of all logs like on SATA and SAS, so opensea-operations was also updated for lengths of standardized logs as best it could be at this time as well.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>