make: Adding musl cross compiler install script and meson cross files
Adding a script to assist with installing musl cross-compilers from musl.cc/#refs for architectures we have built and run these tools on in the past. There are a couple in this list we have not previously targetted, but are likely to work or have minimal issues.
All cross compilers will install to /opt/muslcc at this time and this is where the cross files expect to find the compiler and other tools to build the code.

According to meson docs you can specify multiple cross-files to set variables. This can potentially simplify these files if this is done correctly since so many are basically copy-paste right now.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>