feat: Adding lots of colorful warnings for changing sector sizes (Fast Format)
Due to the number of filed issues both publicly and internally related to issues encountered with setSectorSize or "Fast Format", a lot more warnings are now emitted both at run time and in the help to make sure a user is fully aware that this action can result in a drive becoming unusable due to uncontrollable events such as controller resets while the format runs.
It also needed to be made more clear that the reason this option exists is for compatibility with hardware RAID controllers more than anything else as modern operating systems and filesystems are already 4K aware whether the drives is 4K native or 512 emulation so this is not an option that will provide any additional performance enhancement in most cases.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>