feat: adding NVMe format to openSeaChest_Erase
Adding NVMe format support to openSeaChest_Erase since it can be used to perform a full device erase much like the sanitize commands can work. NVMe format is more likely to be supported by a greater number of devices than sanitize and this was missing from the quickest erase option as well and will improve the user experience further on NVMe devices.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>