feat: Automatically try quick format for Seagate drives after error in FF
If the set sector configuration ext command fails or is interrupted, the drive can enter a indeterminate state and have issues. Seagate has a quick format command that can correct this situation on SATA drives.
I have made this process more automatic with this commit to detect when something has gone wrong and to automatically attempt the recovery process.
I skipped the check for if the command is supported since the support bits in firmware have been removed on some products so it will just attempt the command no matter what happens.
Doing this automatically should help improve the user experience with this option.
Since the quick format recovery is now automatic, the quick format option to run manually has been removed as it is no longer necessary.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>