doc: Updating help to be more clear, especially with device handles
Fixed some help descriptions that had some major issues with clarity and improved help for options that specify a range of LBAs.

Also changed the device handle examples to not use a ? since this is typically to show using a wildcard, but those are not handled in openSeaChest. I went with <#> instead as I thought that would be clear it is expecting some kind of number (in most if not all cases) or some kind of identifier that is specific to the device and stays succinct.

[Seagate/openSeaChest#100]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>