bug: Fixing bugs in transitioning power states for all drive types
Fixing a bug transitioning NVMe power states as it was doing >0 rather than >= 0. 0 is a valid power state on NVMe and should be supported.

For SATA and SAS, the bug was matching the requested transition. This could cause it to transition to an incorrect state. I changed this to strcmp so that it must make an exact match without the possibility of partial matches. This will ensure the correct power state is transitioned to on EPC drives.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>