feat: Initial code changes to support realtek 9210 USB adapter
Checking in initial changes that make supporting the realtek 9210 possible.
This device is pretty different in behavior from previous USB to NVMe chips since it can also support SATA M.2 devices.
Because of this, it required a bit more testing and tweaking to figure out the initial capabilities.
The most peculiar quirk is that

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>