bug: Fixing a hang after starting an erase but not completing it yet
Fixing a hang that can occur when an erase has been started, but is not waiting until the end (not polling). This is because something like sanitize is still running, but the drive is not in a state to accept a read command and the OS is unable to do the filesystem cache update.

[Seagate/openSeaChest#112]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>