doc: Adding a warning about enabling PUIS feature on SAS/SATA HBAs
Adding a warning about enabling the PUIS feature on SAS/SATA HBAs.
Enabling this feature on SATA add-in cards or motherboards is usually fine since the OS is the one managing the feature, however enabling it on a SAS/SATA HBA means that it must be handled by the HBA itself. I do not know of many of these HBAs supporting this feature and if you enabling it the drive may stop showing up when you reboot. This warning in the help should help a little bit by recommending that the HBA's documentation is checked first to prevent this issue.

[Seagate/openSeaChest#111]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>