make: Adding meson build option for ability to set ATA security password in openSeaChest_Security
Adding a meson build option to make it easier to turn on the ability to set an ATA security password.
This option is off by default since it can make a system unbootable, a drive unlockable, or cause other issues that the OS/BIOS/HAB are unable to handle properly.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>