feat: Updating how the NVMe get features option handles requests
Updating when help is shown, how to differentiate between requesting a list of features and a real value.
Also added the ability to specify the feature in hex or decimal.

There is still work to do for features not currently supported in the software (plenty to list) and how to handle unknown or vendor unique features where it is not clear whether these transfer data or not.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>