quick: Adding cast to resolve warning in nvme get feature
Adding a cast to resolve warning about truncation passing a 16bit value into a function expecting an 8bit.
This was not really an issue, but the cast removes the warning and clears up intention that this is ok and expected.
The 16bit value is wider to help determine what a user is requesting the utility to do without using another variable to track for the NVMe getFeature option.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>