bug: Adding "warning" for ATA smart attributes below threshold, but not indicated as pre-fail
Adding a different indicator for attributes at/below the threshold that do not set the pre-failure bit.
These are meant to indicate other warnings such as old age or non-critical issues, so they needed a different way to indicate this to the user.

I also adjusted the hybrid output for attributes 1, 7, 195, and 188 to display the error count since this seems like what people are most interesting in viewing from the raw data.

[Seagate/openSeaChest#108]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>