feat: Removing dumping username to tool banner by default
Removing dumping the username of the user executing the software to the banner by default.
This needs the flag "ENABLE_READ_USERNAME" which matches the same flag in opensea-common.
This was done because a customer reported a security concern due to the low-level linux/unix API to map a UID to a string opening the /etc/passwd file since that is the only mapping available to do this.

Now by default this code only checks if it is running elevated which does not require reading this file and will report the user as either "root"/"admin" or "current user".

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>