feat: Adding option to restore max LBA before erasing a device
Adding a new option that a user can provide to restore the max LBA before starting a drive erasure.
While not necessary for purge capable commands, this is very useful for clear commands to ensure all user data is wiped.
This also helps provide a ready to validate and reuse a drive.

If the restore cannot complete successfully, then an error is generated and the erase will NOT continue until the user has intervened to try to resolve the issue or the option is removed from the command line.

This will handle restoring max LBA regardless of the feature that has reduced it to the best it can.
HPA, AMAC, and DCO are all handled for restoration.
If HPA security is active and an error is encountered, it may be because of this feature and the user will either need to unlock it or power cycle the drive.
The messages printed to the screen help indicate these possible scenarios as well as tips to try connecting the drive a different way or trying a different system.

[Seagate/openSeaChest#52]
[Seagate/openSeaChest#105]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>