feat: Adding detection for when the capacity of a drive has changed, but the adapter does not know
After changing the capacity of a drive, the HBA/bridge/adapter may not know about the change.
So a new function was added so that the user can be prompted to reboot and make sure the adapter knows about the change. This is because in many cases with SATA drives the drive is behind a SATL to translate SCSI info to the host from the ATA drive. This information is often cached, so that cache may be out of date and can only be refreshed by a reboot of the system and thus the adapter.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>