doc: Adding some additional warnings about changing the sector size
Adding/clarifying warnings related to changing the sector size.
There are some OS background activities that a user will not have the ability to control and may interrupt the format.
Also adding a suggestion to use live-USB or other bootable systems to mitigate how many background activities may be running since these are less likely to be polling the drive since it is not yet installed and running on anything yet.

[Seagate/ToolBin#16]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>