lib: Pulling in update to wait up to 1 hour on sector size change
Setting the timeout to an hour to handle the few outlier cases that can happen when changing the sector size with fast format. It is unusal to take more than a few minutes, but checking the lab results there are some cases where it can take longer due to background activity running on the drive when the command is received.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>