feat: PUIS moving to powercontrol and more options to include info and spinup
Moving the PUIS feature code to openSeaChest_PowerControl since this is a power related setting and makes more sense to have there.
openSeaChest_Configure will still accept the command but tells the user to migrate to powercontrol.

I also added an "info" option to see if it is supported, the current state, and whether the spinup command is required or if the drive spins up at the first disk access (r/w) command received.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>