quick: Changing exitcode to int instead of enum
Changing the exit code variable to an int instead of a type of the enum.
Our utilities are setup to use common exit codes and allow for returning utility specific exit codes.
If this is left as an enum type a warning is generated about enum conversion, but this is not really an issue since utility specific exit codes start at a different value, so keeping this variable as an int maintains current expected functionality without a warning.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>