feat: Overhaul on sanitize option
The sanitize option in openSeaChest_Erase has been overhauled to allow specifying "allow unrestricted sanitize exit", "zone no reset", and "no deallocate" for applicable devices/standards.
Ex: zone no reset is only for ZBDs
Ex: no deallocate is only for NVMe devices.

This allows for more verification of erasure or allowing sanitize to run in unrestricted mode, whereas previously only restricted mode was supported.

Also added the ability to specify the number of overwrite passes and the ability to invert the pattern between overwrite passes.

Sanitize info also outputs more information about when a write is required after crypto or block erase on SAS devices, the maximum number of overwrite passes allowed in the standard for a given device, and when freezelock/antifreezelock are supported.

Also added more information about removing the interface cable while sanitize is running in case it takes a long time due to an HBA/OS not handling the sanitize overwrite state and constantly attempting to reset the drive, which slows progress to a crawl.

[Seagate/openSeaChest#102]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>