feat: Updating handling of additional sanitize info and no-deallocate on NVMe
Updated how to output NVMe sanitize unique information as well as how the no-deallocate bit is handled on NVMe devices as an option.

Also handling when a device needs a write after sanitize crypto or block erase (SAS only) to output a message to the user so that they know they need to write new data before reading the device will work as expected.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>