feat: Adding openSeaChest_Raw utility to issue raw commands
Adding openSeaChest_Raw as a utility to issue raw passthrough commands.
It can accept CDBs or TFRs with some info about direction and other flags necessary to build the command and issue it.
These can be helpful for quickly debugging issuing a single specific command when not otherwise available in the other utilities.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>