feat: Merging safe_free updates made internally
Merging in updates to how safe_free is used and adding in use of wrapper functions to allow for type enforcement from the compiler.
Initial conversion to safe_Free with C_CAST(void**) was found to introduce memory leaks in some cases where it converted a struct *** to void**, so this solution fixes those issues and means less casting in the code, but more wrapper functions to allow this to be used safely.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>