ci: Attempting to workaround Clang failures
Clang does not always have a Windows build available for the latest release as it takes time for that to get built and uploaded.
So to work around it it will search for a supported previous release to use instead or exit without an error if one cannot be located.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>