bug: Allowing empty units for various command line options
Many options do not require a unit of measure to be provided but the low-level code was requiring it.
This fixes it so that a unit is not required and allows a utility to choose when a unit is required or what the default behavior should be when the units are empty.

[Seagate/openSeaChest#181]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>