bug: Fixing memory allocation error in openSeaChest_Raw
Fixing a bug where memory will not be allocated properly for issuing a raw command.
Part of the issue is that the device's block size is not populated and part of the issue is unclear command line options and option parsing.
I resolved the option parsing and help text for the options in this commit.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>