feat: Enhancing exit codes when devices don't enumerate properly or get skipped
Changing behavior to better handle when all devices are skipped, have invalid interfaces, or permissions were required when running.
This prevents a zero exit code from being returned by checking if all the devices that were expected to be run were skipped, but if a partial list worked without error, then the exit code is based off the devices that were run.

[Seagate/openSeaChest#43]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>