ci: Removing freebsd riscv64 and openindiana
FreeBSD for riscv64 doesn't support pkg installs right now so we cannot configure necessary tools compatible with the rest of the CI steps.
We can revisit this more later.

OpenIndiana was recently made available and while this action works, it is extremely slow to install all required build tools...like hours long. Not sure if this is an action issue or OpenIndiana issue. Disabling this for now to revisit again later.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>