feat: Merge VMactions CI into meson CI (#252)
* ci: Seeing if I can get omni OS artifacts to upload

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Adding missing argument to --root-user-action to ignore the warning

This VM runs as root so it will install as root.
We can do all the venv work...or just ignore this warning and ignoring this warning is fine on a VM that doesn't preserve its state.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Adding missing step to create omnios artifacts to upload

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Getting more verbose output to troubleshoot tar command

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Passing in DEST_DIR to omnios vm

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Combining VMActions into main meson CI build script

Loading the correct VM Actions into the main meson build script. This will make it easier to generate artifacts and SLSA provenance

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Resolving DragonflyBSD build error and removing old CI steps

Resolving the dragonfly BSD build error (forgot to update install script since it differs from FreeBSD).
Removing separate VMActions build script since the meson.yml can handle it all on its own now, including artifacts and SLSA provenance.

This should be easier to expand to other support VMActions versions and architectures as that project advances and supports more systems and CPUs.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Adding || true after attempting pkg system updates

Making sure that after updating pkg in various OS's that it has || true in case no update is necessary so that CI can continue running other steps.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Updating cirrus to latest FreeBSD versions

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* lib: Pointing to latest libraries

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Adding riscv64 builds for freebsd and openbsd. Adding FreeBSD 15 builds. Adding openindiana build

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Removing freebsd riscv64 and openindiana

FreeBSD for riscv64 doesn't support pkg installs right now so we cannot configure necessary tools compatible with the rest of the CI steps.
We can revisit this more later.

OpenIndiana was recently made available and while this action works, it is extremely slow to install all required build tools...like hours long. Not sure if this is an action issue or OpenIndiana issue. Disabling this for now to revisit again later.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

* ci: Removing FreeBSD 15.0 from CirrusCI since it could not be found

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

---------

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>